requirements.txt
Django==2.2.10
psycopg2==2.8.5
requests==2.23.0
ipython==6.2.1
python-decouple==3.3
Pillow==5.1.0

django-model-utils==3.1.2
django-widget-tweaks==1.4.8